# Allow to write state using compact representation