@crw 2024-10-14
Hi @alexott, just wanted to belatedly say thanks for this contribution. As is probably obvious, this one is on the backburner for now. As I recall, I believe the theory is that the over-the-wire difference should not change too much if using compression during transmission e.g. gzip. Will let you know if this gets back on-track. Thanks!