@zoonage 2025-02-28
@crw this would be nice for some use cases we have too.

We process the state files off the back of them going into S3 so reducing the size reduces our memory requirements and the amount of time spent deserialising data.

I also suspect that some of our TF plans are slow because they are deserialising several GBs of state data that is 40% whitespace